+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 1.0.0
 
 - Initial Release
+
+## 1.0.1
+
+- Added firebase_core library
+- Added Facebook SignIn Method